3rd May

+++ MainForm/CONTACT.cs
@@ -15,8 +15,8 @@ namespace MainForm
         MY_DB mydb = new MY_DB();
         public bool insertContact(int id, string fname, string lname, string groupID, string phone, string email, string address, MemoryStream picture)
         {
-            SqlCommand command = new SqlCommand("INSERT INTO mycontact (id, fname, lname, group_id, phone, email, address, pic)" +
-                "VALUES (@id, @fn, @ln, @gID, @phn, @email, @adr, @pic)", mydb.getConnection);
+            SqlCommand command = new SqlCommand("INSERT INTO mycontact (id, fname, lname, group_id, phone, email, address, pic, userid)" +
+                "VALUES (@id, @fn, @ln, @gID, @phn, @email, @adr, @pic, @uid)", mydb.getConnection);
             command.Parameters.Add("@id", SqlDbType.Int).Value = id;
             command.Parameters.Add("@fn", SqlDbType.VarChar).Value = fname;
             command.Parameters.Add("@ln", SqlDbType.VarChar).Value = lname;
@@ -25,6 +25,7 @@ namespace MainForm
             command.Parameters.Add("@email", SqlDbType.NChar).Value = email;
             command.Parameters.Add("@adr", SqlDbType.Text).Value = address;
             command.Parameters.Add("@pic", SqlDbType.Image).Value = picture.ToArray();
+            command.Parameters.Add("@uid", SqlDbType.Int).Value = Globals.GlobalUserID;
 
             mydb.openConnection();
 
@@ -50,9 +51,10 @@ namespace MainForm
 
         public bool deleteContact(int id)
         {
-            SqlCommand command = new SqlCommand("DELETE FROM mycontact WHERE id = @id", mydb.getConnection);
+            SqlCommand command = new SqlCommand("DELETE FROM mycontact WHERE id = @id AND userid = @uid", mydb.getConnection);
             command.Parameters.Add("@id", SqlDbType.Int).Value = id;
-            
+            command.Parameters.Add("@uid", SqlDbType.Int).Value = Globals.GlobalUserID;
+
             mydb.openConnection();
 
             try
@@ -77,7 +79,7 @@ namespace MainForm
 
         public bool updateContact(int id, string fname, string lname, string groupID, string phone, string email, string address, MemoryStream picture)
         {
-            SqlCommand command = new SqlCommand("UPDATE std SET fname=@fn, lname=@ln, group_id=@gID, phone=@phn, email=@email, address=@adr, picture=@pic WHERE id=@ID", mydb.getConnection);
+            SqlCommand command = new SqlCommand("UPDATE mycontact SET fname=@fn, lname=@ln, group_id=@gID, phone=@phn, email=@email, address=@adr, pic=@pic WHERE id=@ID AND userid = @uid", mydb.getConnection);
             command.Parameters.Add("@ID", SqlDbType.Int).Value = id;
             command.Parameters.Add("@fn", SqlDbType.VarChar).Value = fname;
             command.Parameters.Add("@ln", SqlDbType.VarChar).Value = lname;
@@ -86,6 +88,7 @@ namespace MainForm
             command.Parameters.Add("@email", SqlDbType.NChar).Value = email;
             command.Parameters.Add("@adr", SqlDbType.Text).Value = address;
             command.Parameters.Add("@pic", SqlDbType.Image).Value = picture.ToArray();
+            command.Parameters.Add("@uid", SqlDbType.Int).Value = Globals.GlobalUserID;
 
             mydb.openConnection();
 
@@ -111,6 +114,16 @@ namespace MainForm
 
         public DataTable getContact(SqlCommand command)
         {
+            command.Connection = mydb.getConnection;
+            SqlDataAdapter adapter = new SqlDataAdapter(command);
+            DataTable table = new DataTable();
+            adapter.Fill(table);
+            return table;
+        }
+
+        public DataTable getContactByUserId(SqlCommand command)
+        {
+            command.Parameters.Add("@userid", SqlDbType.Int).Value = Globals.GlobalUserID;
             command.Connection = mydb.getConnection;
             SqlDataAdapter adapter = new SqlDataAdapter(command);
             DataTable table = new DataTable();
